relocate modals to their own directory
